requirements.txt
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0
pillow>=10.0.0
tqdm>=4.65.0
tenacity>=8.2.0
ratelimit>=2.2.1
requests>=2.31.0 